+++ trigerr/__init__.py
@@ -1,6 +1,6 @@
 from .config import config
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"
 
 
 def set_api_key(key):